+++ game/package.json
@@ -1,6 +1,5 @@
 {
   "name": "game",
-  "proxy": "http://127.0.0.1:5000",
   "version": "0.1.0",
   "private": true,
   "type": "module",
